split refactorisation from gem.optimise
